Support override

+++ src/main/java/com/airgap/airgapagent/synchro/work/CopyWork.java
@@ -1,6 +1,7 @@
 package com.airgap.airgapagent.synchro.work;
 
 import com.airgap.airgapagent.synchro.utils.PathInfo;
+import com.airgap.airgapagent.utils.FileUtils;
 
 import java.io.IOException;
 import java.nio.file.Files;
@@ -14,8 +15,12 @@ import java.util.Objects;
  */
 public class CopyWork implements Work {
 
+    private static final int MAX_RETRIES = 100;
 
     private Path target;
+
+    private boolean override = true;
+
 
     public CopyWork() {
         //For jackson
@@ -35,6 +40,14 @@ public class CopyWork implements Work {
 
     public void setTarget(String targetFolder) {
         this.target = Path.of(targetFolder);
+    }
+
+    public boolean isOverride() {
+        return override;
+    }
+
+    public void setOverride(boolean override) {
+        this.override = override;
     }
 
     @Override
@@ -57,6 +70,21 @@ public class CopyWork implements Work {
             Files.createDirectories(parentPath);
         }
         Path targetPath = target.resolve(path.getRelative());
-        Files.copy(path.getOriginalPath(), targetPath, StandardCopyOption.REPLACE_EXISTING);
+        if (!override && Files.exists(targetPath)) {
+            Path expectedPath = null;
+            for (int i = 0; i < MAX_RETRIES; i++) {
+                expectedPath = FileUtils.withRandomTimeStamp(targetPath);
+                if (Files.notExists(expectedPath)) {
+                    break;
+                }
+                expectedPath = null;
+            }
+            if (expectedPath == null) {
+                throw new IllegalStateException("Impossible to find a temporary name");
+            }
+            Files.copy(path.getOriginalPath(), FileUtils.withRandomTimeStamp(expectedPath), StandardCopyOption.REPLACE_EXISTING);
+        } else {
+            Files.copy(path.getOriginalPath(), targetPath, StandardCopyOption.REPLACE_EXISTING);
+        }
     }
 }
